Test Rock
Rock not displasing

+++ Game/ImageObject.h
@@ -5,7 +5,7 @@ class ImageObject
 public:
 	ImageObject(Texture2D texture,float scale, float rotation);
 
-	virtual void draw(int x, int y);
+	virtual void draw(float x, float y);
 	virtual void draw() {};
 private:
 	float rotation, scale;
